+++ spring-xd-analytics/src/main/java/org/springframework/xd/analytics/metrics/core/MetricUtils.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 /**
  * Utility class, primarily to avoid exposing mutable objects beyond the core package.
- *
+ * 
  * For internal use only.
  */
 public final class MetricUtils {
@@ -37,13 +37,13 @@ public final class MetricUtils {
 		return g.reset();
 	}
 
-	public static int[] concatArrays(List<int[]> arrays, int start, int size, int subSize) {
-		int[] counts = new int[size];
+	public static long[] concatArrays(List<long[]> arrays, int start, int size, int subSize) {
+		long[] counts = new long[size];
 
-		for (int i=0; i < arrays.size(); i++) {
-			int[] sub = arrays.get(i);
-			for (int j=0; j < subSize; j++) {
-				int index = i*subSize + j - start; // index in the complete interval
+		for (int i = 0; i < arrays.size(); i++) {
+			long[] sub = arrays.get(i);
+			for (int j = 0; j < subSize; j++) {
+				int index = i * subSize + j - start; // index in the complete interval
 
 				if (index >= counts.length) {
 					break;
